В index.html удаляет лишние подключения script, подключает точку входа, разделяет код main.js на модули.
