+++ index.html
@@ -165,8 +165,6 @@
     </section>
   </template>
 
-  <script src="js/main.js"></script>
-  <script src="js/random.js"></script>
-  <script src="js/utils.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
