--- js/random.js
@@ -1,10 +1,0 @@
-// export function getRandomIntInclusive (min,max) {
-//   if (min > 0 && min < max) {
-//     min = Math.ceil(min);
-//     max = Math.floor(max);
-//     return Math.floor(Math.random() * (max - min + 1)) + min;
-//   }
-//   return NaN;
-// }
-
-// Unexpected token 'export' !!!?
