+++ js/utils.js
@@ -1,5 +1,14 @@
-// export function testCommentLength (string, length) {
-//   return string.length <= length;
-// }
+function getRandomIntInclusive (min,max) {
+  if (min >= 0 && min < max) { //Если ноль не включать, то description: undefined, comments: NaN.
+    min = Math.ceil(min);
+    max = Math.floor(max);
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
+  return NaN;
+}
 
-// Unexpected token 'export' !!!?
+function testCommentLength (string, length) {
+  return string.length <= length;
+}
+
+export { getRandomIntInclusive, testCommentLength};
